+++ system/workspaces.php
@@ -1,0 +1,6 @@
+<?php
+return [
+  'enabled' => true,
+  // default workspace id if not provided
+  'default' => 1,
+];